Feedback Review Adjusted
Great Review, got to know more in details on elastic ip addresses. 
Also from the reference materials, got lot of knowledge on terraform. 
And yes, learnt the best diagrammatic practice by adding the IPv4 CIDR ranges for VPC/Subnet in the diagram. 

As far as putting the Elastic IP cost in the estimated calculation sheet, I don't think it is needed or matters any difference to the overall cost. 
For our usecase, only 1 Elastic IP is needed for the NAT Gateway, at max 2 if we choose the redundant option of having another NAT Gateway in other AZ. 
The Elastic IP cost is free if we use them 100% in a month. We only have to pay if we don't use the Elastic IP for any period, and in our use case, for example, if we don't use any of the or both the elastic IP for the NAT Gateways due to any reason for 100 hrs in a month, then the cost incurs in only 1 USD per month, that's why I don't think it makes any difference to the overall calculation. 
And for the 8 EC2, we don't have to use Elastic IP, as they will be behind application load balancers. 
I hope it is clear. Thank you very much. 

Changelog: Udacity_Diagram_1 is changed with the IPv4 CIDR ranges for VPC and Subnets, and Elastic IP association with NAT Gateway is made visible in the diagram.